+++ src/mkvtoolnix-gui/util/widget.h
@@ -6,7 +6,9 @@
 #include <QDialogButtonBox>
 
 class QComboBox;
+class QFont;
 class QIcon;
+class QLineEdit;
 class QPushButton;
 class QScrollArea;
 class QString;
@@ -37,6 +39,8 @@ void preventScrollingWithoutFocus(QObject *parent);
 void enableChildren(QObject *parent, bool enable);
 
 void addSegmentUIDFromFileToLineEdit(QWidget &parent, QLineEdit &lineEdit, bool append);
+
+QFont defaultUiFont();
 
 }}}